requirements.txt
# Cerca Dai Testi - Dependencies
# Python 3.9+ required

# HTTP requests
requests>=2.28.0

# Environment variables
python-dotenv>=1.0.0

# NLP / Semantic matching
sentence-transformers>=2.2.0
numpy>=1.24.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Type checking (dev)
mypy>=1.0.0

# Linting (dev)
ruff>=0.1.0
